inline layer first throw
