+++ roles/importer/files/importer/checkpointR8x/cp_rule.py
@@ -1,60 +1,52 @@
 from asyncio.log import logger
 from fwo_log import getFwoLogger
 import json
-import cp_const, cpcommon
+import cp_const
 import fwo_const
+import fwo_globals
 from fwo_const import list_delimiter
 from fwo_base import sanitize
 from fwo_exception import ImportRecursionLimitReached
 
-
-def add_section_header_rule_in_json(rulebase, section_name, layer_name, import_id, rule_uid, rule_num, section_header_uids, parent_uid):
-    section_header_uids.append(sanitize(rule_uid))
-    rule = {
-        "control_id":       int(import_id),
-        "rule_num":         int(rule_num),
-        "rulebase_name":    sanitize(layer_name),
-        # rule_ruleid
-        "rule_disabled":    False,
-        "rule_src_neg":     False,
-        "rule_src":         "Any",
-        "rule_src_refs":    sanitize(cp_const.any_obj_uid),
-        "rule_dst_neg":     False,
-        "rule_dst":         "Any",
-        "rule_dst_refs":    sanitize(cp_const.any_obj_uid),
-        "rule_svc_neg":     False,
-        "rule_svc":         "Any",
-        "rule_svc_refs":    sanitize(cp_const.any_obj_uid),
-        "rule_action":      "Accept",
-        "rule_track":       "Log",
-        "rule_installon":   "Policy Targets",
-        "rule_time":        "Any",
-        "rule_implied":      False,
-        # "rule_comment":     None,
-        # rule_name
-        "rule_uid":         sanitize(rule_uid),
-        "rule_head_text":   sanitize(section_name),
-        # rule_from_zone
-        # rule_to_zone
-        # rule_last_change_admin
-        "parent_rule_uid":  sanitize(parent_uid)
-    }
-    rulebase.append(rule)
+uid_to_name_map = {}
 
 
-def add_domain_rule_header_rule_in_json(rulebase, section_name, layer_name, import_id, rule_uid, rule_num, section_header_uids, parent_uid):
-    add_section_header_rule_in_json(rulebase, section_name, layer_name,
-                                    import_id, rule_uid, rule_num, section_header_uids, parent_uid)
+def normalize_rulebases_top_level (full_config, current_import_id, config2import):
+    logger = getFwoLogger()
+    target_rulebase = []
+    rule_num = 0
+    parent_uid=""
+    section_header_uids=[]
 
-
-def resolve_uid_to_name(nw, config2import):
+    # fill uid_to_name_map:
     for nw_obj in config2import['network_objects']:
-        if nw_obj['obj_uid']==nw:
-            return nw_obj['obj_name']
-    return nw
+        uid_to_name_map['obj_uid'] = nw_obj['obj_name']
+
+    rb_range = range(len(full_config['rulebases']))
+    for rb_id in rb_range:
+        # if current_layer_name == args.rulebase:
+        if fwo_globals.debug_level>3:
+            logger.debug("parsing layer " + full_config['rulebases'][rb_id]['layername'])
+
+        # parse access rules
+        rule_num = parse_rulebase(
+            full_config['rulebases'][rb_id], target_rulebase, full_config['rulebases'][rb_id]['layername'], 
+            current_import_id, rule_num, section_header_uids, parent_uid, config2import)
+        # now parse the nat rulebase
+
+        # parse nat rules
+        if len(full_config['nat_rulebases'])>0:
+            if len(full_config['nat_rulebases']) != len(rb_range):
+                logger.warning('get_config - found ' + str(len(full_config['nat_rulebases'])) +
+                    ' nat rulebases and ' +  str(len(rb_range)) + ' access rulebases')
+            else:
+                rule_num = parse_nat_rulebase(
+                    full_config['nat_rulebases'][rb_id], target_rulebase, full_config['rulebases'][rb_id]['layername'], 
+                    current_import_id, rule_num, section_header_uids, parent_uid, config2import)
+    return target_rulebase
 
 
-def parse_single_rule_to_json(src_rule, rulebase, layer_name, import_id, rule_num, parent_uid, config2import, debug_level=0):
+def parse_single_rule(src_rule, rulebase, layer_name, import_id, rule_num, parent_uid, config2import, debug_level=0):
     logger = getFwoLogger()
     # reference to domain rule layer, filling up basic fields
     if 'type' in src_rule and src_rule['type'] != 'place-holder':
@@ -75,8 +67,11 @@ def parse_single_rule_to_json(src_rule, rulebase, layer_name, import_id, rule_nu
                                     src['networks'] + list_delimiter
                         else:  # more than one source
                             for nw in src['networks']:
-                                nw_resolved = resolve_uid_to_name(nw, config2import)
-                                rule_src_name += src["name"] + '@' + nw_resolved + list_delimiter
+                                nw_resolved = resolve_uid_to_name(nw)
+                                if nw_resolved == "":
+                                    rule_src_name += src["name"] + list_delimiter
+                                else:
+                                    rule_src_name += src["name"] + '@' + nw_resolved + list_delimiter
                     else:  # standard network objects as source
                         rule_src_name += src["name"] + list_delimiter
                 else:
@@ -247,67 +242,126 @@ def parse_single_rule_to_json(src_rule, rulebase, layer_name, import_id, rule_nu
                 "last_hit":         sanitize(last_hit)
             }
             rulebase.append(rule)
+            return rule_num + 1
+    return rule_num
 
 
-def parse_rulebase_json(src_rulebase, target_rulebase, layer_name, import_id, rule_num, section_header_uids, parent_uid, config2import, debug_level=0, recursion_level=1):
+def resolve_uid_to_name(nw_obj_uid):
+    if nw_obj_uid in uid_to_name_map:
+        return uid_to_name_map[nw_obj_uid]
+    else:
+        logger = getFwoLogger()
+        logger.warning("could not resolve network object with uid " + nw_obj_uid)
+        return ""
 
-    if (recursion_level > fwo_const.max_recursion_level):
-        raise ImportRecursionLimitReached(
-            "parse_rulebase_json") from None
 
+def insert_section_header_rule(rulebase, section_name, layer_name, import_id, rule_uid, rule_num, section_header_uids, parent_uid):
+    section_header_uids.append(sanitize(rule_uid))
+    rule = {
+        "control_id":       int(import_id),
+        "rule_num":         int(rule_num),
+        "rulebase_name":    sanitize(layer_name),
+        # rule_ruleid
+        "rule_disabled":    False,
+        "rule_src_neg":     False,
+        "rule_src":         "Any",
+        "rule_src_refs":    sanitize(cp_const.any_obj_uid),
+        "rule_dst_neg":     False,
+        "rule_dst":         "Any",
+        "rule_dst_refs":    sanitize(cp_const.any_obj_uid),
+        "rule_svc_neg":     False,
+        "rule_svc":         "Any",
+        "rule_svc_refs":    sanitize(cp_const.any_obj_uid),
+        "rule_action":      "Accept",
+        "rule_track":       "Log",
+        "rule_installon":   "Policy Targets",
+        "rule_time":        "Any",
+        "rule_implied":      False,
+        # "rule_comment":     None,
+        # rule_name
+        "rule_uid":         sanitize(rule_uid),
+        "rule_head_text":   sanitize(section_name),
+        # rule_from_zone
+        # rule_to_zone
+        # rule_last_change_admin
+        "parent_rule_uid":  sanitize(parent_uid)
+    }
+    rulebase.append(rule)
+    return rule_num + 1
+
+
+def add_domain_rule_header_rule(rulebase, section_name, layer_name, import_id, rule_uid, rule_num, section_header_uids, parent_uid):
+    return insert_section_header_rule(rulebase, section_name, layer_name,
+                                    import_id, rule_uid, rule_num, section_header_uids, parent_uid)
+
+
+def check_and_add_section_header(src_rulebase, target_rulebase, layer_name, import_id, rule_num, section_header_uids, parent_uid, config2import, debug_level=0, recursion_level=1):
+    # if current rulebase starts a new section, add section header, but only if it does not exist yet (can happen by chunking a section)
+    if 'type' in src_rulebase and src_rulebase['type'] == 'access-section' and 'uid' in src_rulebase: # and not src_rulebase['uid'] in section_header_uids:
+        section_name = "section without name"
+        if 'name' in src_rulebase:
+            section_name = src_rulebase['name']
+        if 'parent_rule_uid' in src_rulebase:
+            parent_uid = src_rulebase['parent_rule_uid']
+        else:
+            parent_uid = ""
+        rule_num = insert_section_header_rule(target_rulebase, section_name, layer_name, import_id, src_rulebase['uid'], rule_num, section_header_uids, parent_uid)
+        parent_uid = src_rulebase['uid']
+    return rule_num
+
+
+def parse_rulebase(src_rulebase, target_rulebase, layer_name, import_id, rule_num, section_header_uids, parent_uid, config2import, debug_level=0, recursion_level=1):
     logger = getFwoLogger()
-    if 'layerchunks' in src_rulebase:
+    if (recursion_level > fwo_const.max_recursion_level):
+        raise ImportRecursionLimitReached("parse_rulebase") from None
+
+    # parse chunks
+    if 'layerchunks' in src_rulebase:   # found chunks of layers which need to be parsed separately
         for chunk in src_rulebase['layerchunks']:
             if 'rulebase' in chunk:
                 for rules_chunk in chunk['rulebase']:
-                    rule_num = parse_rulebase_json(rules_chunk, target_rulebase, layer_name, import_id, rule_num,
-                                                   section_header_uids, parent_uid, config2import, debug_level=debug_level, recursion_level=recursion_level+1)
+                    rule_num = parse_rulebase(rules_chunk, target_rulebase, layer_name, import_id, rule_num,
+                                                    section_header_uids, parent_uid, config2import, debug_level=debug_level, recursion_level=recursion_level+1)
             else:
-                logger.warning("found no rulebase in chunk:\n" +
-                               json.dumps(chunk, indent=2))
-    else:
-        if 'rulebase' in src_rulebase:
-            # add section header, but only if it does not exist yet (can happen by chunking a section)
-            if src_rulebase['type'] == 'access-section' and not src_rulebase['uid'] in section_header_uids:
-                section_name = "section without name"
-                if 'name' in src_rulebase:
-                    section_name = src_rulebase['name']
-                if 'parent_rule_uid' in src_rulebase:
-                    parent_uid = src_rulebase['parent_rule_uid']
-                else:
-                    parent_uid = ""
-                add_section_header_rule_in_json(target_rulebase, section_name, layer_name,
-                                                import_id, src_rulebase['uid'], rule_num, section_header_uids, parent_uid)
-                rule_num += 1
-                parent_uid = src_rulebase['uid']
-            for rule in src_rulebase['rulebase']:
+                rule_num = parse_rulebase(chunk, target_rulebase, layer_name, import_id, rule_num, section_header_uids, parent_uid, config2import, debug_level=debug_level, recursion_level=recursion_level+1)
+    
+    # # parse rulebase in array (section in layer)
+    # if not 'type' in src_rulebase and len(src_rulebase)==1: # and 0 in src_rulebase:
+    #     rule_num = parse_rulebase(src_rulebase[0], target_rulebase, layer_name, import_id, rule_num, section_header_uids, parent_uid, config2import, debug_level=debug_level, recursion_level=recursion_level)
+        
+    check_and_add_section_header(src_rulebase, target_rulebase, layer_name, import_id, rule_num, section_header_uids, parent_uid, config2import, debug_level=debug_level, recursion_level=recursion_level+1)
+
+    # parse layered rulebase
+    if 'rulebase' in src_rulebase:
+        for rule in src_rulebase['rulebase']:
+            if 'type' in rule:
                 if rule['type'] == 'place-holder':  # add domain rules
                     section_name = ""
                     if 'name' in src_rulebase:
                         section_name = rule['name']
-                    add_domain_rule_header_rule_in_json(
+                    rule_num = add_domain_rule_header_rule(
                         target_rulebase, section_name, layer_name, import_id, rule['uid'], rule_num, section_header_uids, parent_uid)
                 else:  # parse standard sections
-                    parse_single_rule_to_json(
+                    rule_num = parse_single_rule(
                         rule, target_rulebase, layer_name, import_id, rule_num, parent_uid, config2import, debug_level=debug_level)
-                    rule_num += 1
+            if 'rulebase' in rule:  # alsways check if a rule contains another layer
+                rule_num = parse_rulebase(rule, target_rulebase, layer_name, import_id, rule_num, section_header_uids, parent_uid, config2import, debug_level=debug_level, recursion_level=recursion_level+1)
 
-        if src_rulebase['type'] == 'place-holder':  # add domain rules
-            logger.debug('found domain rule ref: ' + src_rulebase['uid'])
-            section_name = ""
-            if 'name' in src_rulebase:
-                section_name = src_rulebase['name']
-            add_domain_rule_header_rule_in_json(
-                target_rulebase, section_name, layer_name, import_id, src_rulebase['uid'], rule_num, section_header_uids, parent_uid)
-            rule_num += 1
-        if 'rule-number' in src_rulebase:   # rulebase is just a single rule
-            parse_single_rule_to_json(
-                src_rulebase, target_rulebase, layer_name, import_id, rule_num, parent_uid, config2import)
-            rule_num += 1
+    if 'type' in src_rulebase and src_rulebase['type'] == 'place-holder':  # add domain rules
+        logger.debug('found domain rule ref: ' + src_rulebase['uid'])
+        section_name = ""
+        if 'name' in src_rulebase:
+            section_name = src_rulebase['name']
+        rule_num = add_domain_rule_header_rule(
+            target_rulebase, section_name, layer_name, import_id, src_rulebase['uid'], rule_num, section_header_uids, parent_uid)
+
+    if 'rule-number' in src_rulebase:   # rulebase is just a single rule
+        rule_num = parse_single_rule(src_rulebase, target_rulebase, layer_name, import_id, rule_num, parent_uid, config2import)
+
     return rule_num
 
 
-def parse_nat_rulebase_json(src_rulebase, target_rulebase, layer_name, import_id, rule_num, section_header_uids, parent_uid, config2import, debug_level=0, recursion_level=1):
+def parse_nat_rulebase(src_rulebase, target_rulebase, layer_name, import_id, rule_num, section_header_uids, parent_uid, config2import, debug_level=0, recursion_level=1):
 
     if (recursion_level > fwo_const.max_recursion_level):
         raise ImportRecursionLimitReached(
@@ -318,39 +372,29 @@ def parse_nat_rulebase_json(src_rulebase, target_rulebase, layer_name, import_id
         for chunk in src_rulebase['nat_rule_chunks']:
             if 'rulebase' in chunk:
                 for rules_chunk in chunk['rulebase']:
-                    rule_num = parse_nat_rulebase_json(rules_chunk, target_rulebase, layer_name, import_id, rule_num,
+                    rule_num = parse_nat_rulebase(rules_chunk, target_rulebase, layer_name, import_id, rule_num,
                                                        section_header_uids, parent_uid, config2import, debug_level=debug_level, recursion_level=recursion_level+1)
             else:
                 logger.warning(
                     "parse_rule: found no rulebase in chunk:\n" + json.dumps(chunk, indent=2))
     else:
         if 'rulebase' in src_rulebase:
-            # add section header, but only if it does not exist yet (can happen by chunking a section)
-            if src_rulebase['type'] == 'access-section' and not src_rulebase['uid'] in section_header_uids:
-                section_name = ""
-                if 'name' in src_rulebase:
-                    section_name = src_rulebase['name']
-                parent_uid = ""
-                add_section_header_rule_in_json(target_rulebase, section_name, layer_name,
-                                                import_id, src_rulebase['uid'], rule_num, section_header_uids, parent_uid)
-                rule_num += 1
-                parent_uid = src_rulebase['uid']
+            check_and_add_section_header(src_rulebase, target_rulebase, layer_name, import_id, rule_num, section_header_uids, parent_uid, config2import, debug_level=debug_level, recursion_level=recursion_level+1)
+
             for rule in src_rulebase['rulebase']:
                 (rule_match, rule_xlate) = parse_nat_rule_transform(rule, rule_num)
-                parse_single_rule_to_json(
+                rule_num = parse_single_rule(
                     rule_match, target_rulebase, layer_name, import_id, rule_num, parent_uid, config2import)
-                parse_single_rule_to_json(
+                parse_single_rule( # do not increase rule_num here
                     rule_xlate, target_rulebase, layer_name, import_id, rule_num, parent_uid, config2import)
-                rule_num += 1
 
-        if 'rule-number' in src_rulebase:   # rulebase is just a single rule
+        if 'rule-number' in src_rulebase:   # rulebase is just a single rule (xlate rules do not count)
             (rule_match, rule_xlate) = parse_nat_rule_transform(
                 src_rulebase, rule_num)
-            parse_single_rule_to_json(
+            rule_num = parse_single_rule(
                 rule_match, target_rulebase, layer_name, import_id, rule_num, parent_uid, config2import)
-            parse_single_rule_to_json(
+            parse_single_rule(  # do not increase rule_num here (xlate rules do not count)
                 rule_xlate, target_rulebase, layer_name, import_id, rule_num, parent_uid, config2import)
-            rule_num += 1
     return rule_num
 
 
@@ -392,3 +436,4 @@ def parse_nat_rule_transform(xlate_rule_in, rule_num):
         'rule_type': 'xlate'
     }
     return (rule_match, rule_xlate)
+
